To delete

+++ App/utils.h
@@ -4,7 +4,7 @@
 #include <sys/types.h>
 #include <string>
 
-void pretty_print_arr(const uint8_t *data, size_t size, size_t max_per_line); 
+void pretty_print_arr(const uint16_t *data, size_t size, size_t max_per_line); 
 int parse_coords(char const *input, struct Coords **coords_arr);
 void print_usage(char *argv[]);
